requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
drf-yasg
django-ckeditor
python-decouple